requirements.txt
pydantic>=2.0.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-telegram-bot>=20.0
python-dotenv>=1.0.0
psutil>=5.9.0
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
